# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AquaStat API",
    "description": "Water-aware infrastructure API for estimating data center water usage and WUE.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/status": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Get Status",
        "operationId": "get_status_api_v1_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/info": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Get Info",
        "operationId": "get_info_api_v1_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIInfoResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIInfoResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "documentation": {
            "type": "string",
            "title": "Documentation"
          },
          "openapi": {
            "type": "string",
            "title": "Openapi"
          },
          "health": {
            "type": "string",
            "title": "Health"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "version",
          "documentation",
          "openapi",
          "health"
        ],
        "title": "APIInfoResponse"
      },
...
      "StatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "environment": {
            "type": "string",
            "title": "Environment"
          },
          "documentation": {
            "type": "string",
            "title": "Documentation"
          },
          "openapi": {
            "type": "string",
            "title": "Openapi"
          }
        },
        "type": "object",
        "required": [
          "status",
          "service",
          "version",
          "environment",
          "documentation",
          "openapi"
        ],
        "title": "StatusResponse"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Railway deployment URL placeholder"
    }
  ]
}